# openapi.yaml
openapi: 3.0.3
info:
  title: AI SmartChat API
  description: |
    API for AI SmartChat application - A smart chat application with AI responses,
    session management, chat history, and background music features.
  version: 1.0.0
  contact:
    name: AI SmartChat Team
servers:
  - url: /api
    description: Local API server
...
paths:
...
  /chat:
    post:
      tags:
        - Chat
      summary: Send message to AI
      description: Send a user message and receive an AI-generated response.
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Successful response from AI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "400":
          description: Bad request - message is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ============================================
  # CHATS - Chat History CRUD
  # ============================================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        response:
          type: string
          description: Optional fallback response message

    # ---------- Session ----------
...
    SendMessageRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: User message to send to AI
        sessionId:
          type: string
          description: Session identifier
        chatId:
          type: string
          description: Chat ID for message context
    SendMessageResponse:
      type: object
      properties:
        response:
          type: string
          description: AI-generated response
        sessionId:
          type: string
          description: Session identifier
        chatId:
          type: string
          description: Chat ID

    # ---------- Music ----------
...